refactor image creation
