+++ convex/_generated/api.d.ts
@@ -18,6 +18,7 @@ import type * as db_admin_see from "../db/admin/see.js";
 import type * as db_chatModels from "../db/chatModels.js";
 import type * as db_endpoints from "../db/endpoints.js";
 import type * as db_imageModels from "../db/imageModels.js";
+import type * as db_images from "../db/images.js";
 import type * as db_messages from "../db/messages.js";
 import type * as db_speech from "../db/speech.js";
 import type * as db_threads from "../db/threads.js";
@@ -33,7 +34,6 @@ import type * as files_ingestImageUrl from "../files/ingestImageUrl.js";
 import type * as files_processUrlContent from "../files/processUrlContent.js";
 import type * as functions from "../functions.js";
 import type * as http from "../http.js";
-import type * as images from "../images.js";
 import type * as inference_assessNsfw from "../inference/assessNsfw.js";
 import type * as inference_captionImage from "../inference/captionImage.js";
 import type * as inference_chat from "../inference/chat.js";
@@ -70,6 +70,7 @@ declare const fullApi: ApiFromModules<{
   "db/chatModels": typeof db_chatModels;
   "db/endpoints": typeof db_endpoints;
   "db/imageModels": typeof db_imageModels;
+  "db/images": typeof db_images;
   "db/messages": typeof db_messages;
   "db/speech": typeof db_speech;
   "db/threads": typeof db_threads;
@@ -85,7 +86,6 @@ declare const fullApi: ApiFromModules<{
   "files/processUrlContent": typeof files_processUrlContent;
   functions: typeof functions;
   http: typeof http;
-  images: typeof images;
   "inference/assessNsfw": typeof inference_assessNsfw;
   "inference/captionImage": typeof inference_captionImage;
   "inference/chat": typeof inference_chat;
