+++ convex/files/ingestImageUrl.ts
@@ -1,15 +1,13 @@
 import { v } from 'convex/values'
 
 import { internal } from '../_generated/api'
-import { getImageModelByResourceKey } from '../db/imageModels'
+import { createImage } from '../db/images'
 import { internalAction, internalMutation } from '../functions'
-import { claimJob, completeJob } from '../jobs'
+import { claimJob, completeJob, handleJobError } from '../jobs'
 import { imageFields } from '../schema'
-import { getTextToImageConfig, insist } from '../shared/utils'
+import { insist } from '../shared/utils'
 
-import type { Ent, MutationCtx } from '../types'
-
-export const claim = internalMutation({
+export const start = internalMutation({
   args: {
     jobId: v.id('jobs'),
   },
@@ -31,19 +29,26 @@ export const run = internalAction({
     jobId: v.id('jobs'),
   },
   handler: async (ctx, args) => {
-    const { job, url, messageId } = await ctx.runMutation(internal.files.ingestImageUrl.claim, args)
+    try {
+      const { job, url, messageId } = await ctx.runMutation(
+        internal.files.ingestImageUrl.start,
+        args,
+      )
 
-    const { fileId, metadata } = await ctx.runAction(internal.lib.sharp.storeImageFromUrl, {
-      url,
-    })
+      const { fileId, metadata } = await ctx.runAction(internal.lib.sharp.storeImageFromUrl, {
+        url,
+      })
 
-    await ctx.runMutation(internal.files.ingestImageUrl.complete, {
-      jobId: job._id,
-      ...metadata,
-      fileId,
-      messageId,
-      sourceUrl: url,
-    })
+      await ctx.runMutation(internal.files.ingestImageUrl.complete, {
+        jobId: job._id,
+        ...metadata,
+        fileId,
+        messageId,
+        sourceUrl: url,
+      })
+    } catch (error) {
+      await handleJobError(ctx, { error, jobId: args.jobId })
+    }
   },
 })
 
@@ -54,37 +59,7 @@ export const complete = internalMutation({
     ...imageFields,
   },
   handler: async (ctx, { jobId, ...args }) => {
-    const message = await ctx.skipRules.table('messages').getX(args.messageId)
-
-    await ctx.table('images').insert({
-      ...args,
-      threadId: message.threadId,
-      userId: message.userId,
-      sourceType: message.role === 'assistant' ? 'textToImage' : 'user',
-      generationData: await getGenerationData(ctx, message),
-    })
-    console.log('[image]', args.sourceUrl)
-
-    if (!message.hasImageContent) {
-      await message.patch({ hasImageContent: true })
-    }
-
+    await createImage(ctx, args)
     await completeJob(ctx, { jobId })
   },
 })
-
-const getGenerationData = async (ctx: MutationCtx, message: Ent<'messages'>) => {
-  const textToImageConfig = getTextToImageConfig(message.inference)
-  if (!textToImageConfig) {
-    return undefined
-  }
-
-  const model = await getImageModelByResourceKey(ctx, textToImageConfig.endpointModelId)
-
-  return {
-    prompt: textToImageConfig.prompt,
-    endpointId: textToImageConfig.endpoint,
-    modelId: textToImageConfig.endpointModelId,
-    modelName: model?.name ?? textToImageConfig.endpointModelId,
-  }
-}
